+++ src/Migration/Manager.php
@@ -17,11 +17,11 @@ final class Manager
 {
     use ParamsCheckerBehavior;
 
-    const TYPE_UP = 'up';
-    const TYPE_DOWN = 'down';
+    public const TYPE_UP = 'up';
+    public const TYPE_DOWN = 'down';
 
-    const TARGET_FIRST = 'first';
-    const TARGET_ALL = 'all';
+    public const TARGET_FIRST = 'first';
+    public const TARGET_ALL = 'all';
 
     private Config $config;