+++ tests/Database/QueryBuilder/MysqlQueryBuilder/CreateTableTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Database\QueryBuilder\MysqlWithJsonQueryBuilder;
+namespace Phoenix\Tests\Database\QueryBuilder\MysqlQueryBuilder;
 
 use Phoenix\Database\Adapter\MysqlAdapter;
 use Phoenix\Database\Element\MigrationTable;
-use Phoenix\Database\QueryBuilder\MysqlWithJsonQueryBuilder;
+use Phoenix\Database\QueryBuilder\MysqlQueryBuilder;
 use Phoenix\Tests\Helpers\Adapter\MysqlCleanupAdapter;
 use Phoenix\Tests\Helpers\Pdo\MysqlPdo;
 use PHPUnit\Framework\TestCase;
@@ -33,7 +33,22 @@ final class CreateTableTest extends TestCase
         $this->assertInstanceOf(MigrationTable::class, $table->addColumn('title', 'string'));
         $this->assertInstanceOf(MigrationTable::class, $table->addColumn('settings', 'json'));
 
-        $queryBuilder = new MysqlWithJsonQueryBuilder($this->adapter);
+        $queryBuilder = new MysqlQueryBuilder($this->adapter);
+        $expectedQueries = [
+            'CREATE TABLE `simple` (`id` int(11) NOT NULL AUTO_INCREMENT,`title` varchar(255) NOT NULL,`settings` text NOT NULL,PRIMARY KEY (`id`)) DEFAULT CHARACTER SET=utf8;'
+        ];
+        $this->assertEquals($expectedQueries, $queryBuilder->createTable($table));
+    }
+
+    public function testSimpleCreateWithJson(): void
+    {
+        $table = new MigrationTable('simple');
+        $table->addPrimary(true);
+        $table->setCharset('utf8');
+        $this->assertInstanceOf(MigrationTable::class, $table->addColumn('title', 'string'));
+        $this->assertInstanceOf(MigrationTable::class, $table->addColumn('settings', 'json'));
+
+        $queryBuilder = new MysqlQueryBuilder($this->adapter, [MysqlQueryBuilder::FEATURE_JSON]);
         $expectedQueries = [
             'CREATE TABLE `simple` (`id` int(11) NOT NULL AUTO_INCREMENT,`title` varchar(255) NOT NULL,`settings` json NOT NULL,PRIMARY KEY (`id`)) DEFAULT CHARACTER SET=utf8;'
         ];