+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### [Unreleased][unreleased]
 
 #### Changed
-- use utf8mb4 as default charset for mysql (fix but BC break)
+- use utf8mb4 as default charset for mysql (fix but BC break, use e.g. `$this->changeCollation('utf8mb4_general_ci')`)
 - dropped support for unsupported PHP versions and added native typehints (BC break)
 - changed autoload to PSR-4
   - moved namespace Dumper to Phoenix\Dumper (BC break)
@@ -12,6 +12,11 @@
 - all classes which can be final are final (BC break if there are some extensions)
 - all methods which can be final are final
 
+#### Removed
+- class MysqlWithJsonQueryBuilder (BC break)
+
+#### Added
+- visibility for constants
 
 
 ### [1.11.1] - 2021-12-07